v0/backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Async and tasks
celery==5.3.6
redis==5.0.1
flower==2.0.1

# YouTube extraction
yt-dlp==2024.3.10

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# AI Transcription
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Language processing and correction
language-tool-python==2.8
spacy==3.7.2
deepmultilingualpunctuation==1.0.1
transformers==4.36.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# HTTP and API
httpx==0.26.0
aiofiles==23.2.1

# Monitoring and logging
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==6.1.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# CORS
fastapi-cors==0.0.6
